`ProcessNode.is_valid_cache`: catch `AttributeError` for unloadable identifier (#5222)
The `is_valid_cache` property is supposed to only raise a `ValueError`
when the function or class represented by the `process_type` identifier
cannot be loaded. However, for a normal Python identifier of an existing
module but with a function or class that does not exist in that module
an `AttributeError` will be thrown which was not caught and converted.