Add the `EnumData` data plugin to easily store `Enum` members
An `Enum` member is represented by three attributes in the database:

 * `name`: the member's name
 * `value`: the member's value
 * `identifier`: the string representation of the enum's identifier

The original enum member can be obtained from the node through the
`get_member` method. This will except if the enum class can no longer be
imported or the stored value is no longer a valid enum value. This can
be the case if the implementation of the class changed since the
original node was stored.

The plugin is named `EnumData`, and not `Enum`, because the latter would
overlap with the Python built-in class `Enum` from the `enum` module.
The downside is that this differs from the naming of other Python base
type analogs, such as `Str` and `Float`, which simply use a capitalized
version of the type for the data plugin class.