scheduler: allow access to computer object
Currently a scheduler knows nothing about its target. While this decoupling is useful for testing, there might be schedulers which require knowledge about their target, especially about things like the base directory of AiiDA calculations on the target. Without this, all but the submit commands are run with the users home dir as their CWD, which may have side-effects (PATH=.:$PATH for example). To make it more reliable, this PR introduces a `cd $working_directory` with the working_directory coming from the computer and being a directory AiiDA anyway assumes full control over it.