Docs: improve the example in the scheduler errors topic
The example was advocating to simply check if any exit status was set on
the node, however, it is probably better to check for specific exit
codes that are known to occurr, such as the OOW and OOM errors that can
be raised by the scheduler.

The example is updated to give both examples and give some advice as to
why the one would be used over the other.