FIX: finalize rename of `verdi database migrate`
`verdi database migrate` was renamed to `verdi storage migrate` in a few places
but a number of commands still mentioned the old command in their output.
In addition, udpating tests to use the new command.

Finally, the redirect from the deprecated `verdi database migrate`
has an import error since it invokes a command that has changed name.
(Not sure why this was not caught by the precommit tests?)