Fix listdir method with pattern for SSH
The `listdir` method of the SSH transport plugin was using
`glob.glob` instead of `self.glob` when a pattern was specified.
This is wrong because it's listing files locally and not on the
remote.
The tests passed because they are running on localhost.
Also, this method is probably never called by AiiDA so it went unnoticed.

This commit fixes the issue. I'm not sure if there is a simple
way to fix the tests - probably we should run them against a
docker container or at least a chroot'ed folder? This might be
another issue to be opened though.

Fixes #5244