CMDLINE: Proposed configure command include profile name
Fixes #4679

When setup computer for not the current profile by `-p <profile>`, the proposed message for
computer configure command does not contain this profile name, which lead to the copy
paste of the proposed message for configure commad not apply to the correct profile.
We read the profile name from ctx.obj and pass it to the proposed configere message.

- Some of unittests in test_computer.py are moved to pytest.