`Process`: raise when `exposed_outputs` gets non-existing `namespace`
Up till now, a call to `exposed_outputs` providing a `namespace` as
argument that doesn't exist, i.e., no outputs were ever exposed in that
namespace, would simply be a no-op. Given that this would often be
because the namespace was accidentally misspelled, the error would go
unnoticed and the process would finish without error message.

To prevent this situation that is difficult to debug, the method will
now raise a `KeyError` if the top-level namespace does not exist.