CLI: add the `verdi storage maintain` command (#4965)
The new command can be used to clean up the storage and optimize its
data usage. The command calls through to utility functions in a newly
added module `aiida.backends.control such that they can also be called
directly through the API if needed.

For now the maintenance only applies to the repository. For this purpose
the abstract repository interface has a new method `maintain`, which
will perform the maintenance operations, if any. It takes a `dry_run`
argument in which case the method should only log what operations would
be taken without actually performing them. The `live` argument can be
used to toggle the execution of operations that should only be done if
they are safe while the repository is in use.

In addition, the `get_info` method was added, which should return key
information about the repository's state and content in the form of a
dictionary. It is added to the output of `verdi storage info`.

The disk object store implementation adds additional keyword arguments
for the `maintain` method which allow the caller to override some of the
defaults that are set when `live=False`.

Instead of the `live` concept, the CLI exposes the `--full` option, which
is essentially the inverse of `live`. The thinking is that `full` is
easier to grasp for a user and by default it should be off as this would
require the AiiDA instance to not be in use and it would take longer.