Fix stalling test
The problem was that the computer that was being duplicated did not
define the `default_mpiprocs_per_machine`, but the `user_input` just
provided enters. Since there was no default, it kept prompting. By
setting the `default_mpiprocs_per_machine` on the computer that is being
duplicated, the problem is fixed. Note that this is done on the
`aiida_localhost` fixture from the `conftest.py`. This allows to remove
the ad-hoc fixture `comp` which was doing exactly the same.