CLI: specify explicit profile in suggested command to configure computer (#5259)
The `verdi computer setup` and `verdi computer duplicate` commands give
a hint of how to configure the computer by specifying the command to
run. The explicit profile was not mentioned, which means that if the
new computer was created with a non-default profile, the suggested
command would be wrong.

Here, the command suggestion always includes the profile to guarantee
that the command is always correct.