First prototype for the profile locking mechanism
From now on aiida will keep track of all processes that request access
to the profile by saving their PIDs inside:

$AIIDA_PATH/access/profile_name/tracked/<process_id>.pid

Before returning control to the client, it will also check that there is
no files of the form:

$AIIDA_PATH/access/profile_name/locking/<process_id>.pid

As this would indicate that the profile is being locked by a process that
requires exclusive access for safety of its operations.

For a process to request such access it will first have to check that there
are no active processes using the profile, so it will look at all files in
the `tracked` folder and compare those to the currently running processes
in the system to check that none is actually active (it will also delete
those outdated tracking files in the process).

The design is as follows:

 - A ProcessData class was defined to store the information relevant to the
   processes.

     - It can be initialized either with a PID (of a process to be looked
       among those currently running in the system) or with a filepath
       (where the data of a previous process was stored, typically either
       in the `locking` or the `tracked` folders). If none of these are
       provided, it will load the info of the currently running process
       that is calling the execution.

     - The class also has a method to write the information to a given
       filepath (typically in the `locking` or the `tracked` folders).

 - An AccessManager class to control the access to a profile.

     - It can be initialized with a profile to which the class will control
       the access to (by default it loads the one currenly in use).

     - It has a couple of internal methods to distribute and modularize the
       different tasks, but the most important externally are:

         - profile_locking_context: a context that can be called to work
           with the profile locked. It will raise LockedProfileError if the
           profile is already locked or LockingProfileError if the profile
           is being accessed by other processes.

         - record_process_access: a method to record the accessing to the
           profile. It is now being called in `load_backend_environment`
           to make sure every process that loads the backend gets recorded.
           It will raise LockedProfileError if the profile is already
           locked.