`Computer`: add the `default_memory_per_machine` property (#5260)
This property allows a user to specify for a `Computer` what default
amount of memory in kB should be requested for a `CalcJob` by the
scheduler. This was already possible through the `max_memory_kb` metadata
option of the `CalcJob`, but in this way, it is possible to specify the
default value that should be used for a computer. If both the computer
specifies a default, but the `max_memory_kb` option is specified, the
latter takes precedence.

Since the `max_memory_kb` already existed and specifies the memory in
kB, the same unit was chosen for the computer property for consistency.

The `verdi computer setup` command adds a new option to define this new
property when creating a new computer.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>