CI: move Jenkins workflow to nightly GHA workflow
Up till now, we have been running a Jenkins server to run certain tests.
The idea was to use this server to run more heavy integration tests that
would clog the CI pipeline, which was running on Travis at the time.

The only test that was added and has been running since are the Reverse
Polish Notation (RPN) workchains. These have uncovered a bug or two, but
it has been rare and the burden of maintaining the Jenkins server is not
neglible. In addition, since then we have moved from Travis to GitHub
Actions as the CI platform, and throughput has not been a problem there.

Therefore, we move the RPN tests to a GHA workflow that runs nightly.
This allows us to decommission the custom Jenkins server while
maintaining the additional coverage that the tests provide, without
overloading the CI by not having them run on every commit.