Add double quotes as an optional to bash script escape
The escape method will escape the string in the quotes. The defalut
implementation is using single quotes to escape the string. However,
single quotes will not eval the spacial shell parameters such as $ and
@, this makes bash script not versatile to set the parameters in
runtime.

The `use_double_quotes` option is added to escape the string in the
double quotes so the spacial parameters are still valid in the shell
script.

Using double escape option into environment variables of job templete
setting address #4836. User can now use double quotes escaping in
calcjob option to escape the environment variables of scheduler.