Alternate implementation
Have `_get_submit_script_environment_variables` include the starting and
end markers, and do not call it in `Scheduler.get_submit_script`, but
instead let the subclasses call it in `_get_submit_script_header`
themselves. This prevents existing external plugins printing the
environment variables twice, and still the logic is provided in a single
place.

This still require external plugins to update to use the new method to
print the environment variables for it to pick up the changes if
`aiida-core` decides to improve the implementation.