Dependencies: explicitly compile `pymatgen` with compatible `numpy`
With the release of `numpy==1.22.0` on December 31 2022, our CI builds
started failing. The reason being that `pymatgen` would get built with a
version of `numpy` that is incompatible with the version of `numpy` that
would be installed in the environment in which the tests are run.

The reason for this is that `pymatgen` does not specify an explicit
version of `numpy` in its build requirements and so will pick the latest
version. But `numpy` only guarantees foreward compatibility with its ABI
and so if compiled with a certain version, the resulting binary won't be
compatible with older `numpy` versions.

In this case explicitly, `pymatgen` was getting built with `numpy==1.22`
but then the tests were run with `1.21.4` which would cause the
exception.

The solution is to ensure `pymatgen` is compiled with a version of
`numpy` that is ABI-compatible with the version that will be used in the
actual test run. The way to do this is to first explicitly install a
version of `numpy` that we need and then install `pymatgen` using the
`--no-build-isolation` flag to force `pip` to not install build
requirements but take them from the existing environment. This is also
why we need to install the `wheel` package explicitly, because this is
also a build requirement of `pymatgen` that no longer will be installed
automatically by `pip` due to `--no-build-isolation`.

Finally, there was a necessary change for the dependencies of Python 3.10
where we update the version of `pymatgen` because older versions could
not be built since the building used CPython code that has been removed
in Python 3.10.