CI: fix failing pre-commit due to `mypy` exception
The pre-commit step was failing due to an exception in `mypy`. This was
caused by a bug in the typing of `sqlalchemy`. Upping the requirement in
the `pre-commit` extra to `sqlalchemy[mypy]~=1.4.29` makes sure that we
install a version that fixes the problem. The `mypy` extra will install
the `sqlalchemy2-stubs` package, so that explicit requirement is removed.
Note that these new versions require `mypy==0.930` so we update that
version.

The update in `mypy` comes with some improvements that allow us to
remove some previous ignore statements and others have to be added or
updated.

In addition, we ensure that we install pinned versions in the workflow
by installing from the requirements file, and then just adding the
additional dependencies from the `pre-commit` extra.