Print warning to stdout on `load_profile` for development versions
The `load_profile` method now calls a `check_version` function. This
checks what version of `aiida-core` is currently installed and if it is
a post release development version, a warning is printed that it should
not be run with production databases. The reason is that we cannot
support automatic data migrations for anything other than released
versions (alpha, beta, and pre-releases are supported).

This function relies on the protocol that development branches will
always use a proper post release development version defined in the
`__version__` attribute of the init file of the `aiida` module. This
means it should be of the form `'2.0.0.post0`.

The choice to place the check of the version in `load_profile` is based
on two requirements:

 * It should be guaranteed to run however the database can be loaded
 * It should ideally only run once and not spam stdout multiple times

The `load_profile` function is the only way to load a profile which is
a necessary step to load any database, therefore this should guarantee
that the version check is always called when it is needed. Another
location could have been `Manager._load_backend` which is also
guaranteed to be called before a database is loaded, however, the
downside here is that this can be called multiple times. this would lead
to the warning being spammed multiple times.

A configuration option `warning.development_version` is added. It is
global only and by default set to true. When set to false, the warning
will no longer be displayed.