CI: move `pytest` configuration to `pyproject.toml` (#5281)
Part of the `pytest` configuration was done through environment
variables in the `.github/workflows/tests.sh`. This is moved to the
`pyproject.toml` where the rest of the configuration is and as such
everything is one place.

The `.coveragerc` configuration file no longer seems necessary since the
configuration is specified directly.