Print warning when incompatible RabbitMQ version is used (#5315)
The lower limit for the `kiwipy` requirement is upped to `>=0.7.5` since
that provides the `server_properties` property on the `RmqCommunicator`.
This allows us to determine the version of the RabbitMQ server that is
configured for the profile. In `load_profile` a check is added that will
print a warning if this version is greater than 3.7.

For RabbitMQ 3.8.15 and up, the default configuration will have a default
timeout on channels of around 30 minutes. This means that tasks that are
not acknowledged within that timelimit will be requeued. If the task was
actually still running with a daemon worker, it can be started by a
second runner simultaneously with all sorts of problems as a consequence.

The breaking change was officially only added in 3.9:

    https://github.com/rabbitmq/rabbitmq-server/commit/80e399212eb35d24969e869d80db60b706e59af3

but it was backported in the v3.8 minor version in v3.8.15.

Cherry-pick: ad28e9b485edb0f74d6ed86d467be4308acdcfd8