`Transport`: store `machine` keyword under `hostname` attribute
A `Transport` is created for a `Computer`, but to keep the transport
class independent of the `Computer` ORM class, it is not passed as an
argument. This means however that, given a transport instance, it was
not possible to know for which `Computer` it was constructed, which is
sometimes necessary for consistency checks.

When the transport is constructed the hostname of the `Computer` is
passed as the `machine` keyword argument, and normally the hostname
should identify the computer uniquely. Note that there is still the
question of for which user the transport is configured, but that is
currently not taken into account.