`RemoteData`: allow passing transport to `_clean` and set extra
Allowing to pass an already open `Transport` to the `_clean` method,
prevents having to open the same transport multiple times when cleaning
many `RemoteData` nodes on the same computer. The downside is that it is
possible for a user to pass a transport that is configured for a
computer that is not the one of the `RemoteData`. A check is added to
prevent this, but it only checks the hostname as it cannot check the
user for which the transport is configured.

In addition to this efficiency update, if the clean operation is
successful, an extra with the key `KEY_EXTRA_CLEANED` is set to `True`.
This is useful to identify `RemoteData` nodes that have been cleaned.

Co-Authored-By: Sebastiaan Huber <mail@sphuber.net>