`verdi calcjob cleanworkdir`: fix filter bug and add filter option
The command did not apply the `AND` clause when both the `-o` and the
`-p` options were provided simultaneously and only the former would be
applied. This is now corrected.

On top of that, a new option `-E/--exit-status` is added that allows to
filter on a specific exit status of the calculation jobs.