`DiskObjectStoreRepositoryBackend`: use container in context manager (#5123)
The `DiskObjectStoreRepositoryBackend` uses a `disk_objectstore.Container`
instance to write and read files. For certain operations, it needs to
create an SqlAlchemy session that will connect to the sqlite database.
It is important that this connection is closed after the operation.

This was not being done, and as a potential result (this has not been
100% confirmed but there is a strong suspicion) we have been seeing the
following error messages in the daemon logs when running:

    sqlite3.ProgrammingError: SQLite objects created in a thread can only
    be used in that same thread. The object was created in thread id
    139722984797952 and this is thread id 139722082379520.

It is not quite clear where the access from multiple threads is coming
from, as the only place threads should be used is in the RabbitMQ
communicator, but that should not have access to the file repository.

Regardless of whether this is the real cause, it is important to close
the connection properly after having used it to allow other processes
to write to the database. To guarantee this, each use of the container
should be done through a context manager. Upon exit, this will guarantee
the `close` method is called on the container, which will properly close
the session. Some quick testing seems to indicate that with these changes
the error messages have disappeared.

Since the container should not be used outside of the class, the public
property is removed. The few tests that were using it, were accessing it
unnecessarily and its use has been removed in favor of a similar method
on a higher level in the interface.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>