Add `ProfileAccessManager` to provide exclusive-access profile locks (#5270)
With the new disk object store repository backend implementation, there
is the need to perform maintenance operations on the repository from
time to time. Some of these operations do not allow other processes to
read from or write to the repository concurrently. Therefore it is
necessary for the maintenance functionality to obtain an exlusive-access
lock on the profile's backend.

To this end the class `aiida.manage.profile_access.ProfileAccessManager`
is added. From now on, when the storage backend of a profile is loaded
it should call the `request_access` method of the manager. This will
cause a PID file to be written to the `ACCESS_CONTROL_DIR` directory of
the given profile. Since this access is non-exclusive, multiple processes
can request access like this concurrently and a PID file will be written
for each, where the command of the process is written to the file. This
information is used in order to be able to determine which PID files
may have gone stale.

When the maintenance operation requires exclusive-access, it should
use the `lock` context manager. If the profile is currently being
accessed or already locked by another process, the lock request will
fail. Otherwise, a lock file is created and this will guarantee that
other processes won't be given a lock or be given normal access.

The choice was made to use a custom implementation for the PID and lock
files instead of using existing libraries such as `filelock`. The reason
for this decision is because the use case is quite specific, where not
only do we have exclusive as well as non-exclusive access, we also
decided that we want to keep a record of the IDs of the processes that
get access, such that when the request for a lock or access is denied,
the error message can provide those PIDs. This will make it easier for
the user to debug which (potentially dead) process is blocking the
profile. The downside of this approach is of course that a custom
implementation is more prone to bugs and cross-platform incompatibility
compared to using a well tested library.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>