Docs: update multi-user info (#5338)
I've had a report that the previous text gave the impression that AiiDA
stores SSH keys in the database. Revising the section, adding some extra
info on the concept of users in AiiDA, and mentioning explicitly that no
sensitive data is stored in the database or file repository.

Also fixes a `pylint` warning for the `phonopy` package.