fix: allow StructureData without specified cell
Since StructureData is supposed to support storing non-periodic
structures, it should also support not specifying a cell.

The default [1,1,1] was both not distinguishable from an actual cell
[1,1,1], and dangerous because it could give the impression that a
structure was ready for a periodic calculation when in fact no cell had
been specified.

The new default `None` clearly signals that no cell was specified.
A check is added that will raise when specifying periodicity without
specifying a cell.