Add extra TemplateCodeInfo class to hold codeinfo for JobTemplate
In the current design, the code_info of calc_job is read from the
code setup and plugin then pass to the job template to create
the bash script. However, job template needs more flexibility to
control the different part of script runline where currently all
the part
  - exec_name from code uuid,
  - code_info.cmdline_params,
  - mpi parameters from computer setting

are stacked together to the job template's code_info. In this PR, the class
`TemplateCodeInfo` is created to handle the elements, where the `code_uuid`
and `withmpi` fields are not used in job script generation.
The code_info of JobTemplate and of `CalcJob` are decoupled
from each other and lead to more flexibility.