`CalcJob`: add option to use double quotes for `environment_variables` (#5349)
The `environment_variables` option of the `CalcJob` process can be used
to have the scheduler plugin define a number of environment variables in
the submit script. By default these are always escaped using single
quotes, just like all other command and declarations in the bash script.

However, this causes problems when the user wants to define an environment
variable that contains a variable itself. The single quotes will cause
it to be interpreted literally by bash instead of de-referencing the
variable itself.

Here we add the `environment_variables_double_quotes` option, which is
a boolean and `False` by default. When set to `True`, the scheduler
plugin should use double quotes to escape the environment variables.