Factories: do not explicitly check type of entry point if `load=False`
The recent release of `setuptools==60.9.0` caused the CI to start
failing. The `CalculationFactory` started raising, when invoked through
the `PluginParamType`, when `load=False` because the loaded entry points
were no longer an instance of `importlib_metadata.EntryPoint`. The
reason for this change is that as of that `setuptools` release, it
started vendoring the `importlib_metadata` package and so it overrides
the type to `setuptools._vendor.importlib_metadata.EntryPoint` for the
loaded entry points.

For reasons unknown, this behavior of the changing of the type of loaded
entry points, only manifested when invoking the CLI through `pytest`.
When calling the failing CLI command directly, it would work just fine.

One solution to this problem would have been to extend the check to also
allow the vendored entry point type of `setuptools` but this seems
fragile. Really, it seems that the original design of checking the type
of the entry point in the case of `load=False` is unnecessary. There is
no reason to suspect that `get_entry_point` should return anything that
does not behave as an entry point if it didn't except. It is better to
be Pythonic here and rely on duck typing. If what is returned by the
factory behaves like an `EntryPoint` instance, regardless of its exact
module and class, then that is all that matters.

The test `tests/plugins/test_entry_point.py` was also explicitly
checking the type of the return value of `get_entry_point` but is
shouldn't as its purpose is just to verify the deprecation warning is
emitted for deprecated entry point names, and so the check is removed.