CI: move time consuming tests to separate workflow that runs nightly (#5354)
The test that runs the Reverse Polish Notation (RPN) workchains and the
migrations tests are relatively time consuming and they should not be
affected by most commits. That's why we can afford to have them just run
nightly instead of on every PR.

A `nightly` GHA workflow is added that runs every day at midnight. It
runs a bash script `tests_nightly.sh` that in turns runs the script
`test_polish_workchains.sh` as well as the unit test using `pytest`. The
latter is called with the option `-m 'nightly'` which will select only
tests that are marked as nightly.

Most tests under `tests/backends/aiida_sqlalchemy/migrations` are marked
as nightly. This is done by applying the `@pytest.mark.nightly` decorator
directly to the test functions. For the tests under the `django_branch`
and `sqlalchemy_branch` this is instead done dynamically through the
`pytest_collection_modifyitems` function in the `conftest.py` module
since having to manually apply this to all test would be too tedious.