Dependencies: update requirement `sphinx~=4.1`
The builds on ReadTheDocs started failing. This was due to release `v2.1`
of `markupsafe` that removed the module `markup_safe`. The version of
`jinja==2.11.3` that was getting installed still imported this. This
version is no longer supported and so they won't be fixing this
incompatibility but instead one should use `jinja>3.0`. This is already
a requirement for `aiida-core`, but the version of `sphinx~=3.2.1` that
we were installing had an upper limit on `jinja` to not install v3
causing the version of jinja to be downgraded. The solution is to update
the version of sphinx which no longer has this limitation of `jinja`.

Finally, we need to require `sphinx~=4.1`, because `v4.0.x` puts an
upper limit `markupsafe<2.0` which is incompatible with `jinja~3.0` but
this upper limit was removed later on.