Dependencies: put upper limit `markupsafe<2.1` (#5371)
The latest release `markupsafe==2.1` breaks `jinja<3.0` and that in
turn breaks `sphinx<4.0`. Upgrading to `sphinx~=4.0` brings its own
problems with a whole host of new warnings that cannot be easily fixed
and so instead we are putting an upper limit on `markupsafe`.

Cherry-pick: 139f7d2507b2d556c22a351c035bd15441e12b35