Tests: do not use `/bin` as folder in `verdi code setup` test (#5376)
Most tests for `verdi code setup` have been defining `/bin/bash` as the
executable since that should exist on most platforms. For remote codes
this poses no problems as it is just the executable name that is stored.
However, for a test storing it as a "local" code, this was causing the
test to take in the order of minutes. The reason is that for local codes
all files of the parent folder of the executable, i.e. `/bin`, are
copied to the repository. To fix the problem, a temporary folder with a
dummy executable is created instead.