`Manager.load_profile`: do not load default profile if other already loaded (#5383)
The `Manager.load_profile` method was loading the default profile if no
explicit profile name was specified, even if another profile was already
loaded. This is violating the note in the docstring that if a profile is
already loaded, not specifying an explicit profile to be loaded should
be a no-op.

This bug was manifesting itself in `verdi` where the `--profile` option
was effectively ignored if a profile other than the default was
specified for a command that uses the `load_dbenv` decorator. This is
because the `ProfileParamType` would correctly load the profile
specified by the option, but then the `load_dbenv` decorator would call
`aiida.cmdline.utils.decorators.load_backend_if_not_loaded` which would
call `Manager.load_profile` without specifying a profile, and so the
default profile would be loaded.