`PsqlDosBackend.maintain`: use `ProfileAccessManager.lock`(#5331)
The `maintain` method of the `PsqlDosBackend` is updated to acquire
exclusive access to a profile by calling `ProfileAccessManager.lock` if
a `full` maintenance is requested. The lock is requested before the call
to `get_unreferenced_keyset` method is called, even though technically
that operation doesn't need it, but it could be confusing to a user if
the maintenance starts working and only after a while refuses access.

The logic is moved from `aiida.storage.control` to the recently created
`PsqlDosBackend`. The same goes for the `get_repository_info` function.
The `verdi storage info` command is updated to now obtain this info from
the storage backend instead of the free functions.