CLI: reinstate the tab-completion of parameter values
We recently removed the `click-completion` dependency that used to
provide the tab-completion functionality for `verdi`, since as of
v8.0 `click` ships itself with an implementation of this feature.

The problem is that the original interface was changed from:

    ParameterType.complete(self, ctx, incomplete)

to:

    ParameterType.shell_complete(self, ctx, param, incomplete)

This change was not added when `click-completion` was dropped and so
value completion was broken, but is reinstated here.